// src/features/home/services/phim.type.ts
export type Banner = {
    hinhAnh: string;
    maBanner: number;
    maPhim:number;
}
export type Phim = {
  maPhim: number
  tenPhim: string
  biDanh: string
  trailer: string
  hinhAnh: string
  moTa: string
  maNhom: string
  ngayKhoiChieu: string // ISO string
  danhGia: number
  hot: boolean
  dangChieu: boolean
  sapChieu: boolean
}